Revised API Backend

+++ Backend/src/main/java/edu/teco/sensordatenbankmanagementsystem/models/Observation.java
@@ -10,10 +10,9 @@ import javax.persistence.Id;
 @Data
 public class Observation {
 
-
     @Id
     @GeneratedValue
-    int id;
+    long id;
 
     @Override
     public String toString() {
